requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
platformdirs==4.3.6
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.6
